requirements.txt
torch
unsloth
ultralytics
streamlit
opencv-python
pillow
datasets
transformers